+++ CHANGELOG.md
@@ -24,7 +24,8 @@ According to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) , the `Unr
 
 ### Added
 
--   [#86](https://github.com/FC4E-WP5/fc4eosc-PIDMR-api/pull/86) - PIDMR-175 Implement GET to retrieve change role request by id as an admin role
+-   [#86](https://github.com/FC4E-WP5/fc4eosc-PIDMR-api/pull/86) - PIDMR-175 Implement GET to retrieve change role request by id as an admin role.
+-   [#99](https://github.com/FC4E-WP5/fc4eosc-PIDMR-api/pull/99) - PIDMR-194 Decode pid URL Parameter to Handle Ampersand (&) Character.
 
 ---
 
@@ -42,7 +43,7 @@ According to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) , the `Unr
 -   [#64](https://github.com/FC4E-WP5/fc4eosc-PIDMR-api/pull/64) - PIDMR-157 Create New Authentication Flow in Keycloak.
 -   [#65](https://github.com/FC4E-WP5/fc4eosc-PIDMR-api/pull/65) - PIDMR-164 Create a new endpoint to get all the promotion requests submitted through our API.
 -   [#66](https://github.com/FC4E-WP5/fc4eosc-PIDMR-api/pull/66) - PIDMR-158 Implement API Endpoint for Assigning User Roles.
--   [#70](https://github.com/FC4E-WP5/fc4eosc-PIDMR-api/pull/70) - PIDMR-169 EMAIL after registration
+-   [#70](https://github.com/FC4E-WP5/fc4eosc-PIDMR-api/pull/70) - PIDMR-169 EMAIL after registration.
 -   [#81](https://github.com/FC4E-WP5/fc4eosc-PIDMR-api/pull/56) - PIDMR-176 Create Endpoint to Retrieve All Role Change Requests by a User.
 -   [#82](https://github.com/FC4E-WP5/fc4eosc-PIDMR-api/pull/82) - PIDMR-177 Remove Assigned Role on Rejected Role Change Request by Admin.